key_file_ddl_log instead of key_file_partition_ddl_log
Monty:

I don't think it's safe to take the key from unique_id. There is no
guarantee that this may not change between different server runs!
Better to create a key_file_ddl_log and use this for all key
operations in ddl recovery.

In practice this is not critical as no one should use performance
schema anyway during recovery!  The few cases of us using recovery
code to rollback rename etc is not important.