MDEV-26301 Split optimization refills temporary table too many times
This an alternative patch that resolves the problem of unnecessary refills
of split derived tables. The patch also improves the estimate of the number
of such refills to lower it in some cases that allows wider usage of split
optimizations.
The patch is not intrusive and does not require any new optimizer switches.