MDEV-29273   Race condition between drop table and closing of table
- This issue caused by race condition between drop thread
and fil_encrypt_thread. fil_encrypt_thread closes the tablespace
if the number of opened files exceeds innodb_open_files.
fil_node_open_file() closes the tablespace which are open and
it doesn't have pending operations. At that time, InnoDB drop
tries to write the redo log for the file delete operation. It throws
the bad file descriptor error.

- When trying to close the file, InnoDB should check whether the
table is going to be dropped.